+++ src/commands/templeosrs/gpearned.ts
@@ -35,7 +35,7 @@ import { templeOverviewTimeAliases } from '../../utils/osrs/inputValidator';
 import {
   templeOverviewTimeValidator,
   formatOverviewTime,
-} from '../../utils/osrs/templeOverviewTime';
+} from '../../utils/osrs/templetime';
 // Anti-spam
 import { antiSpam } from '../../cache/antiSpam';
 // UTILS: Number formatter
