+++ src/utils/osrs/templetime.ts
